% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-cg12-example.R
\name{cg12_example}
\alias{cg12_example}
\title{get a cg12 example data file}
\usage{
cg12_example(file)
}
\arguments{
\item{file}{`<chr>` name of the file}
}
\description{
get a cg12 example data file
}
\examples{
cg12_example("12x96MTP_example.txt")
cg12_example("SR1_results-2.txt")
cg12_example("SR1_results-3.txt")

}
