% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-cg12.R
\name{read_cg12}
\alias{read_cg12}
\title{read results file output from cg12 instrument and output a tidy data frame}
\usage{
read_cg12(file)
}
\arguments{
\item{file}{<chr> path to results file}
}
\description{
read results file output from cg12 instrument and output a tidy data frame
}
\examples{
  # read_cg12("SR1_results.txt")

}
