% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-cg12-example.R
\name{cg12_example}
\alias{cg12_example}
\title{get example file for pkg demo}
\usage{
cg12_example(file)
}
\arguments{
\item{file}{}
}
\description{
get example file for pkg demo
}
\examples{
   cg12_example("SR1_results.txt")
   cg12_example("SR1_results-2.txt")
   cg12_example("SR1_results-3.txt")

}
